# docs/openapi/mobile_api.yaml
openapi: 3.1.0
info:
  title: Ratatoskr Mobile API
  description: RESTful API for Android/iOS mobile clients
  version: 1.0.0
servers:
  - url: https://ratatoskrapi.po4yka.com
    description: Production
  - url: https://staging-ratatoskrapi.po4yka.com
    description: Staging
  - url: http://localhost:8000
    description: Local development
paths:
...
  /v1/signals:
    get:
      tags:
      - Signals
      summary: List Signals
      description: List scored signal candidates for the authenticated user.
      operationId: list_signals_v1_signals_get
      security:
      - HTTPBearer: []
      responses:
        '200':
          description: Signal queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalListResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/signals/sources/{source_id}/active:
    post:
      tags:
      - Signals
      summary: Set Signal Source Active
      description: Enable or pause a signal source if the authenticated user is subscribed to it.
      operationId: set_source_active_v1_signals_sources_source_id_active_post
      security:
      - HTTPBearer: []
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceActiveRequest'
      responses:
        '200':
          description: Source active state updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalSourceActiveResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/signals/{signal_id}/feedback:
    post:
      tags:
      - Signals
      summary: Update Signal Feedback
      description: Write feedback for one signal candidate.
      operationId: update_signal_feedback_v1_signals_signal_id_feedback_post
      security:
      - HTTPBearer: []
      parameters:
      - name: signal_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignalFeedbackRequest'
      responses:
        '200':
          description: Feedback recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalFeedbackResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/signals/topics:
    post:
      tags:
      - Signals
      summary: Upsert Signal Topic
      description: Create or update a single-user topic preference for signal scoring.
      operationId: upsert_topic_v1_signals_topics_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopicPreferenceRequest'
      responses:
        '200':
          description: Topic preference saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalTopicResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
  responses:
...
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Meta:
      type: object
      description: Response metadata for observability and pagination.
      required:
      - correlation_id
      - timestamp
      - version
      - api_version
      properties:
        correlation_id:
          type: string
          description: Request correlation ID for debugging (always present, may be empty string).
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 response timestamp.
        version:
          type: string
          description: App / build version (changes every deploy). Sourced from APP_VERSION env var.
        api_version:
          type: string
          description: |
            API contract semver. Distinct from `version` — bumps only on breaking contract
            changes (response shape, request shape, routing surface). Mobile / CLI clients
            should pin against this value and refuse to upgrade major versions blindly.
        build:
          type: string
          nullable: true
          description: Build identifier (if available).
        pagination:
          $ref: '#/components/schemas/Pagination'
        debug:
          type: object
          nullable: true
          description: Debug information (only in development mode).
    Pagination:
      type: object
      description: Pagination metadata for list responses.
      required:
      - total
      - limit
      - offset
      - hasMore
      properties:
        total:
          type: integer
          description: Total number of items matching the query.
        limit:
          type: integer
          description: Maximum items per page.
        offset:
          type: integer
          description: Current offset from start.
        hasMore:
          type: boolean
          description: Whether more items are available.
        nextCursor:
          type: integer
          nullable: true
          description: Cursor for next page (if using cursor-based pagination).
    BaseSuccessResponse:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          const: true
        meta:
          $ref: '#/components/schemas/Meta'
    ErrorObject:
      type: object
      description: Structured error details for programmatic handling.
      required:
      - code
      - errorType
      - message
      - retryable
      - correlation_id
      properties:
        code:
          type: string
          description: Machine-readable error code for programmatic handling.
          enum:
          # Client errors
          - VALIDATION_ERROR
          - UNAUTHORIZED
          - FORBIDDEN
          - NOT_FOUND
          - CONFLICT
          - RATE_LIMIT_EXCEEDED
          - SESSION_EXPIRED
          # Token/Auth errors
          - TOKEN_EXPIRED
          - TOKEN_INVALID
          - TOKEN_REVOKED
          - TOKEN_WRONG_TYPE
          - REFRESH_RATE_LIMITED
          # Sync errors
          - SYNC_SESSION_EXPIRED
          - SYNC_SESSION_NOT_FOUND
          - SYNC_SESSION_FORBIDDEN
          - SYNC_NO_CHANGES
          - SYNC_CONFLICT
          - SYNC_INVALID_ENTITY
          - SYNC_ENTITY_NOT_FOUND
          # Server errors
          - INTERNAL_ERROR
          - DATABASE_ERROR
          - EXTERNAL_API_ERROR
          - PROCESSING_ERROR
          - AUTH_SERVICE_UNAVAILABLE
          # Configuration errors
          - CONFIGURATION_ERROR
          - FEATURE_DISABLED
        errorType:
          type: string
          description: Error category for client-side handling logic.
          enum:
          - authentication
          - authorization
          - validation
          - not_found
          - conflict
          - rate_limit
          - external_service
          - sync
          - internal
          - configuration
        message:
          type: string
          description: Human-readable error message.
        retryable:
          type: boolean
          description: Whether the client should retry the request.
          default: false
        details:
          type: object
          nullable: true
          description: Additional error context (field errors, resource IDs, etc.).
        correlation_id:
          type: string
          description: Request correlation ID for debugging (always present).
        retry_after:
          type: integer
          nullable: true
          description: Seconds to wait before retrying (for rate limit errors).
    ErrorResponse:
      type: object
      required:
      - success
      - error
      - meta
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
...
    SignalFeedbackRequest:
      type: object
      required:
      - action
      properties:
        action:
          type: string
          enum:
          - like
          - dislike
          - skip
          - hide_source
          - queue
          - boost_topic
    SourceActiveRequest:
      type: object
      required:
      - is_active
      properties:
        is_active:
          type: boolean
    TopicPreferenceRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        description:
          type: string
          nullable: true
          maxLength: 500
        weight:
          type: number
          minimum: 0
          maximum: 5
          default: 1
    SignalItem:
      type: object
      additionalProperties: true
      properties:
        id:
          type: integer
        status:
          type: string
        heuristic_score:
          type: number
          nullable: true
        llm_score:
          type: number
          nullable: true
        final_score:
          type: number
          nullable: true
        filter_stage:
          type: string
          nullable: true
        feed_item_title:
          type: string
          nullable: true
        feed_item_url:
          type: string
          nullable: true
        source_kind:
          type: string
          nullable: true
        source_title:
          type: string
          nullable: true
        topic_name:
          type: string
          nullable: true
...
    SignalListResponseEnvelope:
      allOf:
      - $ref: '#/components/schemas/BaseSuccessResponse'
      - type: object
        required:
        - data
        properties:
          data:
            type: object
            required:
            - signals
            properties:
              signals:
                type: array
                items:
                  $ref: '#/components/schemas/SignalItem'
...
    SignalFeedbackResponseEnvelope:
      allOf:
      - $ref: '#/components/schemas/BaseSuccessResponse'
      - type: object
        required:
        - data
        properties:
          data:
            type: object
            properties:
              updated:
                type: boolean
    SignalSourceActiveResponseEnvelope:
      allOf:
      - $ref: '#/components/schemas/BaseSuccessResponse'
      - type: object
        required:
        - data
        properties:
          data:
            type: object
            properties:
              updated:
                type: boolean
              is_active:
                type: boolean
    SignalTopicResponseEnvelope:
      allOf:
      - $ref: '#/components/schemas/BaseSuccessResponse'
      - type: object
        required:
        - data
        properties:
          data:
            type: object
            properties:
              topic:
                type: object
                additionalProperties: true
...